browserプラグインを小分けしWebWorker側と共用できるものを利用するよう修正。 nako3/pluginの取り込むを共通の取り込むに統一。
