+++ src/plugin_browser_hotkey.js
@@ -1,0 +1,22 @@
+const hotkeys = require('hotkeys-js')
+
+module.exports = {
+  // @ホットキー
+  'ホットキー登録': { // @ホットキーKEYにEVENTを登録する // @ほっときーとうろく
+    type: 'func',
+    josi: [['に', 'で'], ['を']],
+    fn: function (key, fname, sys) {
+      hotkeys(key, function (event, handler) {
+        event.preventDefault()
+        sys.__v1[fname]()
+      })
+    }
+  },
+  'ホットキー解除': { // @ホットキーKEYを解除する // @ほっときーかいじょ
+    type: 'func',
+    josi: [['を', 'の']],
+    fn: function (key) {
+      hotkeys.unbind(key)
+    }
+  }
+}
