+++ src/plugin_browser_html.js
@@ -1,0 +1,13 @@
+module.exports = {
+  // @HTML操作
+  'HTML変換': { // @文字列をHTMLに変換して返す // @HTMLへんかん
+    type: 'func',
+    josi: [['を']],
+    fn: function (text) {
+      return String(text)
+        .replace(/&/g, '&amp;')
+        .replace(/>/g, '&gt;')
+        .replace(/</g, '&lt;')
+    }
+  }
+}
